requirements.txt
# Core Discord
discord.py>=2.3.0
py-cord>=2.4.0

# AI/ML Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0

# Image Processing
Pillow>=10.0.0

# Ollama Integration
ollama>=0.1.0

# Hugging Face
huggingface-hub>=0.16.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0

# Optional but recommended
numpy>=1.24.0
scipy>=1.10.0
